Simulators

SymbolicPlanners.jl extends the PDDL.Simulator interface to enable simulation of plan or policy Solutions:

PDDL.Simulator — Method

(sim::Simulator)(sol::Solution, domain::Domain, state::State, [spec])

Simulates the execution of sol in a PDDL domain starting from an initial state. A goal specification spec may be provided to serve as a terminating condition for the simulation, or to specify cost or reward information.

PDDL.Simulator — Method

(sim::Simulator)(domain::Domain, state::State, actions, spec)

Simulates the execution of actions in a PDDL domain starting from an initial state. The goal specification spec serves as a terminating condition for the simulation, and may also specify cost or reward information.

The following simulators are can be used to simulate solutions:

PDDL.EndStateSimulator — Type

EndStateSimulator(max_steps::Union{Int,Nothing} = nothing)

Simulator that returns the end state of simulation.

PDDL.StateRecorder — Type

StateRecorder(max_steps::Union{Int,Nothing} = nothing)

Simulator that records the state trajectory, including the start state.

SymbolicPlanners.StateActionRecorder — Type

StateActionRecorder(max_steps::Union{Int,Nothing} = nothing)

Simulator that records the state-action trajectory, including the start state.

SymbolicPlanners.RewardAccumulator — Type

RewardAccumulator(max_steps::Union{Int,Nothing} = nothing)

Simulator that returns accumulated reward. A specification should be provided when using this simulator.